# Speed up `Matrix4x4.CreateScale`